Added basic check detection and forced moving.
When valid moves are calculated, the game calculates whether the King
would be in check after the move. If so, the move is not considered
valid. Valid moves therefore only contains moves that do not leave the
King in check.
This is inefficient, as most moves are guaranteed to not leave the king
in check. See notes in Todo.txt for improvements.